Eliminated compilation warning.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@19097 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/cssaver/cssaver.cpp
@@ -824,7 +824,7 @@ bool csSaver::SaveSectorLights(iSector *s, iDocumentNode *parent)
           float intensity = cross->GetIntensityFactor();
           csRef<iDocumentNode>intensityNode = CreateNode(haloNode, "intensity");
           intensityNode->CreateNodeBefore(CS_NODE_TEXT)
-            ->SetValueAsInt(intensity);
+            ->SetValueAsInt((int)intensity);
 
           float crossfact = cross->GetCrossFactor();
           csRef<iDocumentNode>crossfactNode = CreateNode(haloNode, "cross");
